getSetPointBounds
